pyproject.toml
[tool.poetry]
name = "orchestration-agent"
version = "0.1.0"
description = ""
authors = ["KennyVaneetvelde <kenny@inosta.be>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
instructor = {version = "1.9.2", extras = ["cerebras-cloud-sdk"]}
pydantic = ">=2.10.3,<3.0.0"
sympy = "^1.13.3"
python-dotenv = ">=1.0.1,<2.0.0"
openai = ">=1.35.12,<2.0.0"
atomic-agents = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
